Migrate row actions
