+++ apps/upskii/src/app/[locale]/(dashboard)/[wsId]/courses/page.tsx
@@ -4,8 +4,8 @@ import { CoursePagination } from './course-pagination';
 import { CustomDataTable } from '@/components/custom-data-table';
 import { createClient } from '@tuturuuu/supabase/next/server';
 import { WorkspaceCourse } from '@tuturuuu/types/db';
-import { CourseForm } from '@tuturuuu/ui/custom/education/courses/course-form';
 import { CourseCardView } from '@tuturuuu/ui/custom/education/courses/course-card-view';
+import { CourseForm } from '@tuturuuu/ui/custom/education/courses/course-form';
 import FeatureSummary from '@tuturuuu/ui/custom/feature-summary';
 import { ViewToggle } from '@tuturuuu/ui/custom/view-toggle';
 import { Separator } from '@tuturuuu/ui/separator';
